+++ test/plugin/scenarios/jedis-3.1.x-plus-scenario/src/main/java/org/apache/skywalking/apm/testcase/jedis/controller/RedisPipelineCommandExecutor.java
@@ -36,6 +36,15 @@ public class RedisPipelineCommandExecutor implements AutoCloseable {
         pipeline.syncAndReturnAll();
     }
 
+    public void pipelineDiscard() {
+        Pipeline pipeline = jedis.pipelined();
+        pipeline.multi();
+        pipeline.hset("a", "a", "a");
+        pipeline.hget("a", "a");
+        pipeline.hdel("a", "a");
+        pipeline.discard();
+    }
+
     public void close() throws Exception {
         jedis.close();
     }